Moved pages and styles into source and fixed a few issues in stories.
