+++ .storybook/preview.js
@@ -1,4 +1,4 @@
-import "../styles/globals.css"
+import "../src/styles/globals.css"
 
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
