+++ src/stories/pages/about.stories.tsx
@@ -1,5 +1,5 @@
 import DefaultLayout from "components/templates/layouts/default-layout";
-import About from "../../../pages/about";
+import About from "../../pages/about";
 
 const storyConfig = {
   title: "Design System/Pages/About"
