+++ src/stories/pages/home.stories.tsx
@@ -1,5 +1,5 @@
 import DefaultLayout from "components/templates/layouts/default-layout";
-import Home from "../../../pages";
+import Home from "../../pages";
 
 const storyConfig = {
   title: "Design System/Pages/Home"
